separate dummyItems.
